Move ViewImageDrawer to /ImageLibrary; update Pendo IDs
