+++ packages/manager/src/features/Images/ImagesLanding/v2/constants.ts
@@ -4,4 +4,5 @@ export const DEFAULT_PAGE_SIZES = [25, 50, 75, 100];
 export const VIEW_SHARED_IMAGE_DETAILS_DRAWER_PENDO_IDS = {
   xButton: 'Images Library Shared View-X button',
   closeButton: 'Images Library Shared View-Close',
+  copyImageIdIcon: 'Images Library Shared View-Copy ID',
 };
